HBASE-8651 Result of integer multiplication cast to long in HRegionFileSystem#sleepBeforeRetry() (Samar)
git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1493229 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/HRegionFileSystem.java
@@ -998,6 +998,6 @@ public class HRegionFileSystem {
       return;
     }
     LOG.debug(msg + ", sleeping " + baseSleepBeforeRetries + " times " + sleepMultiplier);
-    Threads.sleep(baseSleepBeforeRetries * sleepMultiplier);
+    Threads.sleep((long)baseSleepBeforeRetries * sleepMultiplier);
   }
 }
